refactor(tests): remove obsolete test modules from Redis, string, set, sorted set, and admin primitives; consolidate tests into dedicated files for improved organization and maintainability
