+++ crates/adapter/src/redis/primitives/set.rs
@@ -551,220 +551,3 @@ impl RedisSet {
         )
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use redis::pipe;
-    use std::sync::{Arc, Mutex};
-
-    // Create test connection for compilation validation
-    fn create_test_connection() -> Arc<Mutex<redis::Connection>> {
-        // Test client without Redis server dependency
-        let redis_url =
-            std::env::var("REDIS_URL").unwrap_or_else(|_| "redis://127.0.0.1:6379".to_string());
-        let client = redis::Client::open(redis_url).unwrap_or_else(|_| {
-            redis::Client::open("redis://localhost:6379").expect("Creating test client")
-        });
-
-        // Connection object for compilation
-        match client.get_connection() {
-            Ok(conn) => Arc::new(Mutex::new(conn)),
-            Err(_) => {
-                // Connection unavailable in test environment
-                let client =
-                    redis::Client::open("redis://localhost:6379").expect("Creating test client");
-                let conn = client.get_connection().unwrap_or_else(|_| {
-                    panic!("This test is only for compilation and is marked as ignored")
-                });
-                Arc::new(Mutex::new(conn))
-            }
-        }
-    }
-
-    #[test]
-    #[ignore = "Compilation test only"]
-    fn test_compile_operations() {
-        // Compilation verification test
-        let conn = create_test_connection();
-        let redis_set = RedisSet::new(conn);
-
-        // Verify compilation
-        let _sadd_cmd = redis_set.sadd("test_set", &["member1", "member2"]);
-        let _smembers_cmd = redis_set.smembers("test_set");
-        let _srem_cmd = redis_set.srem("test_set", &["member1"]);
-        let _scard_cmd = redis_set.scard("test_set");
-        let _sismember_cmd = redis_set.sismember("test_set", "member1");
-        let _srandmember_cmd = redis_set.srandmember("test_set");
-        let _spop_cmd = redis_set.spop("test_set");
-        let _smove_cmd = redis_set.smove("set1", "set2", "member1");
-        let _sinter_cmd = redis_set.sinter(&["set1", "set2"]);
-        let _sunion_cmd = redis_set.sunion(&["set1", "set2"]);
-        let _sdiff_cmd = redis_set.sdiff(&["set1", "set2"]);
-        let _sinterstore_cmd = redis_set.sinterstore("dest", &["set1", "set2"]);
-        let _sunionstore_cmd = redis_set.sunionstore("dest", &["set1", "set2"]);
-        let _sdiffstore_cmd = redis_set.sdiffstore("dest", &["set1", "set2"]);
-    }
-
-    #[test]
-    #[ignore = "Compilation test only"]
-    fn test_pipeline_methods() {
-        // Test that pipelines can be used directly with cmd()
-        let mut pipeline = pipe();
-
-        let _pipe_ref1 = pipeline
-            .cmd("SADD")
-            .arg("set1")
-            .arg("member1")
-            .arg("member2");
-        let _pipe_ref2 = pipeline.cmd("SMEMBERS").arg("set1");
-        let _pipe_ref3 = pipeline.cmd("SCARD").arg("set1");
-    }
-
-    #[test]
-    #[ignore = "Compilation test only"]
-    fn test_batch_operations() {
-        let conn = create_test_connection();
-        let redis_set = RedisSet::new(conn);
-
-        // Test data for batch operations
-        let set_data = vec![
-            ("set1", vec!["member1", "member2", "member3"]),
-            ("set2", vec!["member2", "member3", "member4"]),
-            ("set3", vec!["member1", "member4", "member5"]),
-        ];
-
-        // Verify method compilation
-        let _ = redis_set.sadd_many(set_data);
-
-        // Test batch remove
-        let remove_data = vec![("set1", vec!["member1"]), ("set2", vec!["member2"])];
-        let _ = redis_set.srem_many(remove_data);
-
-        // Test batch get members
-        let keys = vec!["set1", "set2", "set3"];
-        let _ = redis_set.smembers_many(keys);
-
-        // Test batch check membership
-        let membership_checks = vec![("set1", "member1"), ("set2", "member2")];
-        let _ = redis_set.sismember_many(membership_checks);
-
-        // Test batch get cardinalities
-        let set_keys = vec!["set1", "set2", "set3"];
-        let _ = redis_set.scard_many(set_keys);
-
-        // Test batch delete
-        let expired_keys = vec!["old_set1", "old_set2"];
-        let _ = redis_set.del_many(expired_keys);
-    }
-
-    #[test]
-    #[ignore = "Compilation test only"]
-    fn test_lua_scripts() {
-        let conn = create_test_connection();
-        let _redis_set = RedisSet::new(conn);
-
-        // Create some example scripts
-        let _script = RedisSet::create_script("return redis.call('SCARD', KEYS[1])");
-        let _add_script = RedisSet::add_and_get_cardinality_script();
-
-        // Test pipeline integration with scripts
-        let mut pipe = redis::pipe();
-        _redis_set.add_script_to_pipeline(
-            &mut pipe,
-            script_constants::PING_SCRIPT,
-            vec!["set1"],
-            vec!["new_member"],
-        );
-    }
-
-    #[test]
-    #[ignore = "Compilation test only"]
-    fn test_transaction() {
-        let conn = create_test_connection();
-        let _redis_set = RedisSet::new(conn);
-
-        // Compilation verification only
-    }
-
-    // Integration tests require Redis instance setup.
-}
-
-/// Examples of how to use RedisSet with various features
-///
-/// These examples demonstrate how to use RedisSet's features
-/// in real-world scenarios.
-#[cfg(test)]
-mod examples {
-    use super::*;
-
-    #[test]
-    #[ignore = "Demonstration only"]
-    fn example_patterns() {
-        // Example connection setup
-        let redis_url =
-            std::env::var("REDIS_URL").unwrap_or_else(|_| "redis://127.0.0.1:6379".to_string());
-        let client = redis::Client::open(redis_url).unwrap_or_else(|_| {
-            redis::Client::open("redis://localhost:6379").expect("Creating example client")
-        });
-
-        // Example connection object
-        let conn =
-            Arc::new(Mutex::new(client.get_connection().unwrap_or_else(|_| {
-                panic!("Demonstration example - ignored test")
-            })));
-
-        let redis_set = RedisSet::new(conn);
-
-        // Create a script for demonstration
-        let _add_script = RedisSet::create_script("return redis.call('SADD', KEYS[1], ARGV[1])");
-
-        // Example 1: Pipeline with multiple set operations
-        let _: Result<(usize, Vec<String>), redis::RedisError> = redis_set.with_pipeline(|pipe| {
-            pipe.cmd("SADD")
-                .arg("set1")
-                .arg("member1")
-                .arg("member2")
-                .cmd("SMEMBERS")
-                .arg("set1")
-                .cmd("SCARD")
-                .arg("set1")
-        });
-
-        // Example 2: Transaction with multiple set operations
-        let _: Result<(usize, usize), redis::RedisError> = redis_set.transaction(|pipe| {
-            pipe.cmd("SADD")
-                .arg("tx:set1")
-                .arg("member1")
-                .cmd("SADD")
-                .arg("tx:set2")
-                .arg("member2")
-                .cmd("EXPIRE")
-                .arg("tx:set1")
-                .arg(3600)
-        });
-
-        // Example 3: Using scripts in pipelines
-        let _: Result<(usize, Vec<String>), redis::RedisError> = redis_set.with_pipeline(|pipe| {
-            redis_set.add_script_to_pipeline(
-                pipe,
-                script_constants::PING_SCRIPT,
-                vec!["set1"],
-                vec!["new_member"],
-            );
-
-            pipe.cmd("SMEMBERS").arg("set1")
-        });
-
-        // Example 4: Batch operations
-        let _ = redis_set.sadd_many(vec![
-            ("batch:set1", vec!["member1", "member2"]),
-            ("batch:set2", vec!["member2", "member3"]),
-        ]);
-
-        // Example 5: Set operations
-        let _ = redis_set.sinter(&["set1", "set2"]);
-        let _ = redis_set.sunion(&["set1", "set2"]);
-        let _ = redis_set.sdiff(&["set1", "set2"]);
-    }
-}
