+++ crates/dbx-api/src/tests/route_tests/mod.rs
@@ -1,8 +1,7 @@
 //! Route handler test modules
 
-// TODO: Add route tests when route handlers are refactored
-// /// API key route tests
-// mod api_key_route_tests;
+/// API key route tests
+mod api_key_route_tests;
 
-// /// Role management route tests
-// mod role_route_tests;
+/// Role management route tests
+mod role_route_tests;
